pyproject.toml
[tool.poetry]
name = "microblog"
version = "0.1.0"
description = ""
authors = ["samiribrh <samiribrahimov2277@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
flask-wtf = "^1.2.1"
flake8 = "^7.1.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
